Move utility closer to its users
